refactor: rename folder models to types
