+++ frontend/src/app/shared/types/auth/auth-state.ts
