+++ frontend/src/app/shared/types/auth/password-reset.ts
