+++ frontend/src/app/shared/types/auth/sign-in-response.ts
