+++ frontend/src/app/shared/types/user-details.ts
